requirements_render.txt
Flask>=2.0.0
gunicorn>=20.1.0
Werkzeug>=2.0.0
firebase-admin>=5.0.0
google-generativeai>=0.3.0
google-cloud-speech>=2.0.0
google-cloud-translate>=3.0.0
tiktoken>=0.9.0
requests>=2.28.0
python-dotenv>=0.20.0
